requirements.txt
requests
pandas
flask
langchain
langchain-openai
langchain-community
chromadb
sentence-transformers
python-dotenv
